Convert plugin to Gradle
