--- plugin/src/io/twitchmc/config.yml
@@ -1,2 +1,0 @@
-api_domain: "https://twitchmc.io/api"
-server_id: ""
