--- plugin/src/io/twitchmc/plugin.yml
@@ -1,10 +1,0 @@
-name: TwitchMC
-version: 1.0
-author: AlthalusAvan
-main: io.twitchmc.TwitchMC
-api-version: 1.18
-commands:
-  twitchmc:
-    description: Use various TwitchMC commands
-    usage: "/twitchmc [register] [code]"
-    permission: twitchmc.all
