+++ plugin/src/main/resources/config.yml
@@ -1,0 +1,2 @@
+api_domain: "https://twitchmc.io/api"
+server_id: ""
